Create custom_functions.php
this lines solves the problem to have bizztheme and WooCommerce to play perfekt togheter.
This time i have problem with the layout of my product page. the faulty one has separated the page in 2 equal sized parts, and the correct has at 1to 3 ration.
Cause and consequesns, it seems to have something to do with the order i installed the part. Perhaps you can help me correct this issue